SCON-452: Split LicensePanel into LicenseSection and UpsellSection
